Consolidate to performanceIsomorphic, update sha.js types, remove type dependency from odsp-driver
