+++ common/lib/common-utils/src/test/types/tsconfig.json
@@ -1,19 +1,14 @@
 {
-    "extends": "@fluidframework/build-common/ts-common-config.json",
-    "exclude": [
-        "dist",
-        "node_modules"
-    ],
-    "compilerOptions": {
-        "rootDir": "./",
-        "outDir": "../../../dist/test/types",
-    },
-    "include": [
-        "./**/*"
-    ],
-    "references": [
-        {
-            "path": "../../.."
-        }
-    ]
+	"extends": "@fluidframework/build-common/ts-common-config.json",
+	"exclude": ["dist", "node_modules"],
+	"compilerOptions": {
+		"rootDir": "./",
+		"outDir": "../../../dist/test/types",
+	},
+	"include": ["./**/*"],
+	"references": [
+		{
+			"path": "../../..",
+		},
+	],
 }
