+++ common/lib/common-utils/tsconfig.json
@@ -5,9 +5,6 @@
 		"rootDir": "./src",
 		"outDir": "./dist",
 		"baseUrl": ".",
-		"paths": {
-			"perf_hooks": ["types/perf_hooks.d.ts"],
-		},
 	},
 	"include": ["src/**/*"],
 	"exclude": ["src/test/**/*"],
