+++ packages/common/client-utils/package.json
@@ -89,7 +89,7 @@
 		"@types/node": "^16.18.38",
 		"@types/puppeteer": "1.3.0",
 		"@types/rewire": "^2.5.28",
-		"@types/sha.js": "^2.4.0",
+		"@types/sha.js": "^2.4.4",
 		"@types/sinon": "^7.0.13",
 		"c8": "^7.7.1",
 		"concurrently": "^8.2.1",
