+++ packages/common/client-utils/src/hashFileNode.ts
@@ -8,6 +8,8 @@ import sha1 from "sha.js/sha1";
 // eslint-disable-next-line import/no-internal-modules
 import sha256 from "sha.js/sha256";
 
+// import { sha1, sha256 } from "sha.js";
+
 import type { IsoBuffer } from "./bufferNode";
 
 /**
