+++ packages/common/client-utils/src/indexNode.ts
@@ -11,4 +11,4 @@ export {
 	Uint8ArrayToString,
 } from "./bufferNode";
 export { gitHashFile, hashFile } from "./hashFileNode";
-export { performance } from "./performanceNode";
+export { performance } from "./performanceIsomorphic";
