+++ packages/drivers/odsp-driver/package.json
@@ -86,7 +86,6 @@
 		"@types/mocha": "^9.1.1",
 		"@types/node": "^16.18.38",
 		"@types/node-fetch": "^2.5.10",
-		"@types/sha.js": "^2.4.0",
 		"@types/sinon": "^7.0.13",
 		"c8": "^7.7.1",
 		"cross-env": "^7.0.3",
